requirements.txt
adal==1.2.7
alembic==1.9.4
asgiref==3.6.0
asttokens==2.2.1
azure-common==1.1.28
azure-core==1.26.3
azure-functions==1.13.3
azure-identity==1.12.0
azure-mgmt-billing==6.0.0
azure-mgmt-consumption==10.0.0
azure-mgmt-core==1.3.2
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-keyvault==1.1.0
azure-mgmt-loganalytics==0.2.0
azure-mgmt-logic==3.0.0
azure-mgmt-machinelearningcompute==0.4.1
azure-mgmt-managementgroups==0.1.0
azure-mgmt-managementpartner==0.1.1
azure-mgmt-maps==0.1.0
azure-mgmt-marketplaceordering==0.1.0
azure-mgmt-media==1.0.1
azure-mgmt-monitor==0.5.2
azure-mgmt-msi==0.2.0
azure-mgmt-network==2.7.0
azure-mgmt-notificationhubs==2.1.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==0.1.0
azure-mgmt-powerbiembedded==2.0.0
azure-mgmt-rdbms==1.9.0
azure-mgmt-recoveryservices==0.3.0
azure-mgmt-recoveryservicesbackup==0.3.0
azure-mgmt-redis==5.0.0
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==0.2.1
azure-mgmt-resource==2.2.0
azure-mgmt-scheduler==2.0.0
azure-mgmt-search==2.1.0
azure-mgmt-servicebus==0.5.3
azure-mgmt-servicefabric==0.2.0
azure-mgmt-signalr==0.1.1
azure-mgmt-sql==0.9.1
azure-mgmt-storage==2.0.0
azure-mgmt-subscription==0.2.0
azure-mgmt-trafficmanager==0.50.0
azure-mgmt-web==0.35.0
azure-nspkg==3.0.2
backcall==0.2.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.2
cryptography==40.0.1
debugpy==1.6.6
decorator==5.1.1
Django==4.1.7
et-xmlfile==1.1.0
executing==1.2.0
Flask==2.2.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
ipykernel==6.21.2
ipython==8.10.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.0.3
jupyter_core==5.2.0
Mako==1.2.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
momoapi==1.0.1
msal==1.21.0
msal-extensions==1.0.0
msrest==0.7.1
msrestazure==0.6.4
nest-asyncio==1.5.6
oauthlib==3.2.2
openpyxl==3.0.10
packaging==23.0
parso==0.8.3
phonenumbers==8.13.6
pickleshare==0.7.5
platformdirs==3.0.0
portalocker==1.7.1
prompt-toolkit==3.0.37
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pymssql==2.2.7
pyodbc==4.0.35
python-dateutil==2.8.2
pyzmq==25.0.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==2.0.4
sqlparse==0.4.3
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.2
